pyproject.toml
[tool.poetry]
name = "blog-engine"
version = "0.1.0"
description = "Engine for blog using Django"
authors = ["Marcelo Machado <mjoelmachadojr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
django = "^5.0.4"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
blue = "^0.9.1"
isort = "^5.13.2"
ruff = "^0.3.5"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff && isort --check . --diff'
format = 'blue . && isort .'
run = 'python engine/manage.py runserver'
user = 'python engine/manage.py createuser'
superuser = 'python engine/manage.py createsuperuser'
makemigration = 'python engine/manage.py makemigrations'
migrate = 'python engine/manage.py migrate'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
